support multiple package declarations & add basic parser integration test (#6)
This adds a simple integration test for the Scala code parser, with 9
Scala files pulled from several open source projects (our own fsq.io,
the Scala compiler itself, and Spark). Expected parser output for each
of them is included here as json files.

In adding these, I hit an error around the multiple package declarations
in the scalac source files. This seemed pretty trivial to support, so
that is included here as well.