remove scala.reflect.runtime.universe from default package map (#11)
Internally, we list dependencies for `scala-reflect` and
`scala-compiler` explicitly in `deps` lists. This is possibly
unnecessary and just a holdover from our Pants days however, so we do
not want to force this behavior on all users.

The default package map used for 3rdparty jar dependencies is more or
less impossible to override, so I am removing a hard-coded resolve
mapping for `scala.reflect.runtime.universe` here. If desired, a `#
gazelle:resolve scala scala.reflect.runtime.universe
@maven//:org_scala_lang_scala_reflect` directive can serve as a
replacement

Fixes #10